% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainability.R
\name{PDPmatrix}
\alias{PDPmatrix}
\title{Scatterplot matrix of 2D partial dependence plots}
\usage{
PDPmatrix(model, x, vnames, depth = 21, alpha = 2/3, ...)
}
\arguments{
\item{model}{A model with corresponding predict function that returns numeric values.}

\item{x}{Data frame.}

\item{vnames}{Character vector of the variable set for which the patial dependence function is to be computed.}

\item{depth}{Integer specifiying the number colours in the heat map.}

\item{alpha}{Numeric value for alpha blending of the points in the scatter plot.}

\item{...}{Further arguments to be passed to the \code{predict()} method of the \code{model}.}
}
\description{
Creates a scatterplot matrix of 2D partial dependence plots.
}
\examples{
\dontrun{
library(pdp)
library(randomForest)
data(boston)
set.seed(42)
boston.rf <- randomForest(cmedv ~ ., data = boston)
PDPmatrix(boston.rf, boston, vnames = c("lstat", "rm", "lon", "nox"))
}

}
\references{
\itemize{
    \item Szepannek, G. (2019): How Much Can We See? A Note on Quantifying Explainability of Machine Learning Models,
          \href{https://arxiv.org/abs/1910.13376}{\emph{arXiv:1910.13376 [stat.ML]}}.
  }
}
\author{
\email{gero.szepannek@web.de}
}
